@@ -29,9 +29,9 @@ structlog = ">=21.1,<23.0"
 dunamai = "^1.15.0"
 flake8 = "^5.0.4"
 flake8-isort = "^6.0.0"
-ipython = "^8.9.0"
+ipython = "^8.10.0"
 isort = "^5.12.0"
-konch = "^4.2.2"
+konch = "^4.4.0"
 mypy = "^0.991"
 pdbpp = "^0.10.3"
 pre-commit = "^2.21.0"
@@ -43,7 +43,7 @@ pytest-xdist = "^3.2.0"
 sphinx = "^5.3.0"
 tox = "^4.4.5"
 black = "^22.12.0"
-hypothesis = "^6.67.1"
+hypothesis = "^6.68.1"
 mutmut = "^2.4.3"
 
 [tool.mypy]
